Made Java script in modules - Ok
