+++ index.html
@@ -104,6 +104,9 @@
         nomodule
         src="https://unpkg.com/ionicons@7.1.0/dist/ionicons/ionicons.js"></script>
       
-      <script src="./public/js/controller.js"></script>
+      <script 
+        type="module"
+        src="./public/js/controller.js">
+      </script>
   </body>
 </html>
